% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dj94functions.R
\name{blocks}
\alias{blocks}
\alias{blocks_dj94}
\title{Blocks function}
\usage{
blocks(x, t, h)

blocks_dj94(x)
}
\arguments{
\item{x}{numeric. See description.}

\item{t}{numeric. See description.}

\item{h}{numeric. See description.}
}
\description{
The blocks function from Donoho and Johnstone (1994),
}
\details{
\deqn{
f(x) = \sum h_j K(x - t_j); K(t) = \frac{1 + sgn(x)}{2}
}{
 f(x) = sum( h_j K(x - t_j); K(t) = (1 + sgn(x)) / 2 )
}
The \code{blocks} function is this general case.
The \code{blocks_dj94} function uses the values of \code{t_j} and \code{h_j}
in the original paper
\deqn{
(t_j) = (0.1, 0.13, 0.15, 0.23, 0.25, 0.40, 0.44, 0.65, 0.76, 0.78, 0.81)
}
and
\deqn{
(h_j) = (4, -5, 3, -4, 5, -4.2, 2.1, 4.3, -3.1, 2.1, -4.2)
}
The original paper used the range (0, 1).
}
\examples{
x <- seq(0, 1, length = 2048)
plot(x, blocks_dj94(x), type = "l")
}
\references{
Donoho, David L, and Johnstone, Iain M. (1994) \dQuote{Ideal spatial adaptation by wavelet shrinkage.}
\emph{Biometrika} \strong{81.3}, 425-55.
}

