% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CP6}
\alias{CP6}
\title{CP6 Data}
\format{A \code{data.frame} with \Sexpr[stage=build]{ncol(bsdlm::CP6)} variables, and \Sexpr[stage=build]{nrow(bsdlm::CP6)} observations.}
\usage{
CP6
}
\description{
Monthly total sales, in monetary terms, of tobacco and related products marketed by a U.K. company from January 1955 to December 1959.
}
\references{
West, Mike and Harrison, Jeff. (1997) \emph{Bayesian Forecasting and Dynamic Models, 2nd ed.}. p. 369-370.
}
\keyword{datasets}

