% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{divisas}
\alias{divisas}
\title{Exchange rate data for\Sexpr{ncol(bsdlm::divisas)} currencies}
\format{A \code{\link[zoo]{zoo}} object with \Sexpr{ncol(bsdlm::divisas)} variables and \Sexpr{nrow(bsdlm::divisas)}  observations.}
\source{
\url{http://www.jstatsoft.org/v39/i02/supp/2}
}
\usage{
divisas
}
\description{
Daily observations of exchage rates against the U.S. dollar (USD) for the period January 4, 1988 to August 15, 1997.
}
\references{
Tusell, Fernando. (2011) Kalman Filtering in R. \emph{Journal of Statistical Software} \strong{39.2}. \url{http://www.jstatsoft.org/v39/i02}
}
\keyword{datasets}

