% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dlm_stats.R
\name{dlm_stats}
\alias{dlm_stats}
\title{Calculate misc stats for a DLM}
\usage{
dlm_stats(y, mod)
}
\arguments{
\item{y}{The data. 'y' can be a vector, a matrix, a univariate or multivariate time series.}

\item{mod}{An object of class 'dlm'}
}
\value{
A \code{list} with elements
\describe{
\item{filter}{An object of class \code{\link[dlm:dlmFilter]{dlmFiltered}}}
\item{smooth}{The output of \code{\link[dlm]{dlmSmooth}}}
\item{yrep}{The output of \code{\link[dlmextras]{dlmYrep}} using smoothed draws.}
\item{nu}{The output of \code{\link[dlmextras]{dlmNu}}}
\item{omega}{The output of \code{\link[dlmextras]{dlmOmega}}}
\item{loglik}{The output of \code{\link[dlmextras]{dlmLogLikObs}}}
}
}
\description{
Calculate misc stats for a DLM
}

