% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dj94functions.R
\name{heavisine}
\alias{heavisine}
\alias{heavisine_dj94}
\title{HeaviSine Function}
\usage{
heavisine(x, a, b, c)

heavisine_dj94(x)
}
\arguments{
\item{x}{numeric. See description.}

\item{a}{numeric. See description.}

\item{b}{numeric. See description.}

\item{c}{numeric. See description.}
}
\description{
The HeaviSine function fom Donoho and Johnstone (1994),
\deqn{
f(x) = a \sin a \pi x - sgn (x - b) - sgn(c - x)
} {
f(x) = a * sin(a * pi * x) - sgn(x - b) - sgn(c - x)
}
The function \code{heavisine} is the general case.
The function \code{heavisine_dj94} is the case used in the paper: \eqn{a = 4}, \eqn{b = 0.3}, and \eqn{c = 0.72}
}
\examples{
x <- seq(0, 1, length.out = 2048)
plot(x, heavisine_dj94(x), type = "l")
}
\references{
Donoho, David L, and Johnstone, Iain M. (1994) \dQuote{Ideal spatial adaptation by wavelet shrinkage.}
\emph{Biometrika} \strong{81.3}, 425-55.
}
\seealso{
Other DonohoJohnstone1994: \code{\link{bumps}},
  \code{\link{bumps_dj94}}; \code{\link{doppler}},
  \code{\link{doppler_dj94}}
}

