% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dhp.R
\name{dhorseshoe}
\alias{dhorseshoe}
\alias{rhorseshoe}
\title{The Horseshoe Distribution}
\usage{
dhorseshoe(x, location = 0, scale = 1)

rhorseshoe(n, location = 0, scale = 1)
}
\arguments{
\item{x}{\code{numeric} vector of quantiles.}

\item{location}{\code{numeric}. Location parameter.}

\item{scale}{\code{numeric}. Scale parameter, greater than zero.}

\item{n}{\code{integer} number of observations.}
}
\value{
\code{numeric} vector of density values.
}
\description{
Density function and random generation for the horseshoe distribution.
}
\section{Details}{


There is no exact analytical form of the horseshoe distribution.
It is defined as a scale mixture of normal distributions.

\deqn{
p(x) \sim \mathcal{N}(0, \lambda^2)
}{p(x | lambda, tau) ~ N(0, \lambda^2)}
\deqn{
\lambda \sim \mathcal{C}^{+}(0, \tau)
}{lambda ~ C+(0, tau)}
where \eqn{C^{+}(0, s)}{C+(0, s)} is the half-Cauchy distribution with support on the positive real numbers and
scale \eqn{s}.

However, Theorem 1 of Carvallo et al. (2010) provide bounds on the univariate
density function.  For \eqn{x \neq 0},
\deqn{
\frac{K}{2} \log \left(1 + \frac{4}{x^2} \right) < p(x) < K \log \left( 1 + \frac{2}{x^2}\right)
}{(K / 2) * log(1 + 4 / x^2) < p(x) < K * log(1 + 2 / x^2)}
where \eqn{K = 1 / \sqrt{2 \pi^3}}{K = 1 / sqrt(2 * pi^3)}.
}
\examples{
x <- seq(-4, 4, by = 0.01)
y <- dhorseshoe(x)
plot(x, y, ylim= c(0, 1), type = "l")
z <- rhorseshoe(5000)
plot(density(z), xlim = c(-4, 4))
}
\references{
Carvalho, Carlos M., Nicholas G. Polson, and James G. Scott (2010). \dQuote{The horseshoe estimator for sparse
signals}. \emph{Biometrika}, \strong{97.2}, pp. 465-480. \url{http://biomet.oxfordjournals.org/content/97/2/465.abstract}.

Carvalho, Carlos M., Nicholas G. Polson, and James G. Scott (2009). \dQuote{Handling Sparsity via the Horseshoe}.
\emph{Journal of Machine Learning and Research: Workshop and Conference Proceedings} \strong{5}, pp. 73-80.
}

