% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dhp.R
\name{dhorseshoeplus}
\alias{dhorseshoeplus}
\alias{rhorseshoeplus}
\title{The Horseshoe+ Distribution}
\usage{
dhorseshoeplus(x, location = 0, scale = 1)

rhorseshoeplus(n, location = 0, scale = 1)
}
\arguments{
\item{x}{\code{numeric} vector of quantiles.}

\item{location}{\code{numeric}. Location parameter.}

\item{scale}{\code{numeric}. Scale parameter, greater than zero.}

\item{n}{\code{integer} number of observations.}
}
\value{
\code{numeric} vector of density values.
}
\description{
Density function and random generation for the horseshoe+ prior distribution.
}
\section{Details}{


There is no exact analytical form of the horseshoe+ prior distribution.
It is defined as a scale mixture of normal distributions:
\deqn{
p(x | \lambda, \tau, \eta) \sim \mathcal{N}(0, \lambda^2)
}{p(x | lambda, tau, eta) ~ N(0, lambda^2)}
\deqn{
\lambda \sim \mathcal{C}^{+}(0, \tau \eta)
}{lambda ~ C^+(0, tau eta)}
\deqn{
\eta \sim \mathcal{C}^+(0, 1)
}{eta ~ C^+(0, 1)}
and \eqn{C^{+}(0, s)} is the half-Cauchy distribution with support on the positive real numbers and scale \eqn{s}.

However, Theorem 2 of Bhadra et al. (2015) provide bounds on the univariate  density function.  For \eqn{x \neq 0},
\deqn{
\frac{1}{\pi^2 \sqrt{2 \pi}} \log \left(1 + \frac{4}{x^2} \right) < p(x) < \frac{1}{\pi^2 |x|}
}{1 / (pi^2 * sqrt(2 * pi)) log(1 + 4 / x^2) < p(x) < 1 / (pi^2 * |x|)}
}
\examples{
x <- seq(-4, 4, by = 0.01)
y <- dhorseshoeplus(x)
plot(x, y, ylim= c(0, 1), type = "l")
z <- rhorseshoeplus(5000)
plot(density(z), xlim = c(-4, 4))
}
\references{
Bhadra, Anindya and Datta, Jyotishka and Polson, Nicholas G. and Willard, Brandon. (2015). \dQuote{The Horseshoe+ Estimator of Ultra-Sparse Signals}.
arXiv:1502.00560 [math.ST]. \url{http://arxiv.org/abs/1502.00560v2}.
}

